Replace redundant submit tests with query tests in test_transformation_request.py to match resource name
